+++ src/test/kotlin/TestSnippet.kt
@@ -82,7 +82,7 @@ int adder(int first, int second) {
     }
 })
 
-fun haveParseErrorOnLine(line: Long) = object : Matcher<SnippetParsingFailed> {
+fun haveParseErrorOnLine(line: Int) = object : Matcher<SnippetParsingFailed> {
     override fun test(value: SnippetParsingFailed): Result {
         return Result(value.errors.any { it.location.line == line },
                 "should have parse error on line $line",
